+++ pkgs/dart_mcp/CHANGELOG.md
@@ -10,6 +10,8 @@
 - Fixed a bug where the examples would only connect to a server with the latest
   protocol version.
 - Added validation for Schema extension.
+- Fixed an issue where getting the type of a Schema with a null type would
+  throw.
 
 ## 0.2.0